feat: 최단 경로 조회 서비스 구현

+++ src/main/java/subway/domain/FareCalculator.java
@@ -3,11 +3,11 @@ package subway.domain;
 public class FareCalculator {
 
     private static final int STANDARD_FARE = 1_250;
-    private static final int STANDARD_LOWER_DISTANCE = 10;
-    private static final int STANDARD_UPPER_DISTANCE = 50;
+    private static final Long STANDARD_LOWER_DISTANCE = 10L;
+    private static final Long STANDARD_UPPER_DISTANCE = 50L;
 
 
-    public static int calculate(int distance) {
+    public static int calculate(Long distance) {
         if (distance > STANDARD_LOWER_DISTANCE && distance <= STANDARD_UPPER_DISTANCE) {
             return STANDARD_FARE + calculateOverFareWhenNotExceedUpperDistance(distance - STANDARD_LOWER_DISTANCE);
         }
@@ -19,11 +19,11 @@ public class FareCalculator {
         return STANDARD_FARE;
     }
 
-    private static int calculateOverFareWhenNotExceedUpperDistance(int overDistance) {
-        return (((overDistance - 1) / 5) + 1) * 100;
+    private static int calculateOverFareWhenNotExceedUpperDistance(Long overDistance) {
+        return (int) ((((overDistance - 1) / 5) + 1) * 100);
     }
 
-    private static int calculateOverFareWhenExceedUpperDistance(int overDistance) {
-        return (((overDistance - 1) / 8) + 1) * 100;
+    private static int calculateOverFareWhenExceedUpperDistance(Long overDistance) {
+        return (int) ((((overDistance - 1) / 8) + 1) * 100);
     }
 }
